[1657] Extract out BaseCLI from CoursesEditorCLI
