+++ bin/mcb
@@ -23,6 +23,7 @@ require 'mcb/azure'
 require 'mcb/config'
 require 'mcb/course_show'
 require 'mcb/courses_editor'
+require 'mcb/base_cli'
 require 'mcb/courses_editor_cli'
 require 'mcb/grant_access_wizard'
 require 'mcb/revoke_access_wizard'
