+++ lib/mcb/courses_editor_cli.rb
@@ -1,7 +1,7 @@
 module MCB
-  class CoursesEditorCLI
+  class CoursesEditorCLI < BaseCLI
     def initialize(provider)
-      @cli = HighLine.new
+      super()
       @provider = provider
     end
 
@@ -88,43 +88,6 @@ module MCB
       @cli.ask("Course code?  ", ->(str) { str.upcase }) do |q|
         q.whitespace = :strip_and_collapse
         q.validate = /\S+/
-      end
-    end
-
-    def multiselect(initial_items:, possible_items:)
-      selected_items = initial_items
-      finished = false
-      until finished do
-        @cli.choose do |menu|
-          menu.choice("continue") { finished = true }
-          define_choices_for_each_possible_item(
-            menu: menu,
-            selected_items: selected_items,
-            possible_items: possible_items
-          )
-        end
-      end
-      selected_items
-    end
-
-    def ask_multiple_choice(prompt:, choices:, default: nil)
-      @cli.choose do |menu|
-        menu.prompt = prompt + "  "
-        menu.choice("exit") { nil }
-        menu.choices(*choices)
-        menu.default = default if default.present?
-      end
-    end
-
-  private
-
-    def define_choices_for_each_possible_item(menu:, selected_items:, possible_items:)
-      possible_items.sort_by(&:to_s).each do |item|
-        if item.in?(selected_items)
-          menu.choice("[x] #{item}") { selected_items.delete(item) }
-        else
-          menu.choice("[ ] #{item}") { selected_items << item }
-        end
       end
     end
   end
